v2.0 Mejora de todos los CRUD del sistema

+++ huellas_sanas/apps/Transaccion/models.py
@@ -11,7 +11,7 @@ class Producto(models.Model):
     nombre = models.CharField(max_length=100)
     marca = models.CharField(max_length=100)
     descripcion = models.TextField()
-    precio = models.DecimalField(max_digits=10, decimal_places=2)
+    precio = models.PositiveIntegerField() 
     cantidad_stock = models.PositiveIntegerField()
 
     def __str__(self):
@@ -21,7 +21,7 @@ class Producto(models.Model):
 class Servicio(models.Model):
     nombre = models.CharField(max_length=100)
     descripcion = models.TextField()
-    precio = models.DecimalField(max_digits=10, decimal_places=2)
+    precio = models.PositiveIntegerField() 
 
     def __str__(self):
         return self.nombre
